feat: Move notification modal in Header folder and make notification api file
