+++ components/Header/NotificationModal.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import ModalFrame from './ModalFrame';
 import styles from './NotificationModal.module.scss';
 
+import ModalFrame from '@components/ModalFrame';
 import { MODAL_NAMES, useModal } from '@contexts/ModalContext';
 
 export default function NotificationModal() {
