+++ lib/apis/notification.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+import { apiEndPoint } from './endpoint';
+
+export const getFollowRequests = () => {};
+
+export const getScheduleRequests = () => {};
